Add documentation and enhance codes
Introduced comments for providing clear explanation to each class, method and properties. This will guide the readers to understand the reasoning and purpose behind the each line of code. Additionally, the code has been enhanced for better efficiency by making adjustments in method parameters, updating some access modifiers and modifying certain method's functionalities.

For example, an optional boolean parameter `fromUpdater` was added to the `ResourceUpdater` delegate in buffer resource class to ascertain whether the resource had been previously updated. `ResourceInfo` constructor was changed from public to internal to restrict its accessibility. Consequentially, the calling of the constructor from outside the class or assembly is not possible anymore, which allows greater control. One of the glaring functionality modifications is in the 'DoubleBuffer' class where the '_swapEffect' field is now readonly.

This commit would reduce complexity and improve the readability of the system.